[FEAT] 저장한 조합 조회, 특정 조합의 상세 조회 API 구현

+++ src/main/java/com/refit/app/domain/combination/dto/CombinationProductDto.java
@@ -13,8 +13,8 @@ public class CombinationProductDto {
     private Long productId;
     private String productName;
     private String brandName;
-    private int price;
-    private int discountRate;
-    private int discountedPrice;
+    private Long price;
+    private Integer discountRate;
+    private Long discountedPrice;
     private String thumbnailUrl;
 }
